\name{input_2016}
\docType{data}
\alias{input_2016}
\title{Input Data for the 2016 ACIC Competition}
\description{
  Input data used in the 2016 Atlantic Causal Inference Competition, taken from the Collaborative Perinatal Project.
}
\usage{input_2016}
\format{A data frame consisting of 4802 observations and 58 covariates. The columns have been de-identified
        from their original source, but correspond to possible confounders, instruments, and uncorrelated
        variables from a hypothetical twin study on the impact of birthweight on IQ.}
\details{
  The variable in the original CPP are:
  \itemize{
    \item \code{mom_age}
    \item \code{mar_status}
    \item \code{mom_cigs_per_day}
    \item \code{mom_years_smoked}
    \item \code{mom_height}
    \item \code{mom_weight_prior}
    \item \code{mom_num_cardio_cond}
    \item \code{mom_num_pulm_cond}
    \item \code{mom_num_hema_cond}
    \item \code{mom_num_endocrine_cond}
    \item \code{mom_num_veneral_cond}
    \item \code{mom_num_urin_cond}
    \item \code{mom_num_gyne_cond}
    \item \code{mom_num_neur_cond}
    \item \code{mom_num_obst_compl}
    \item \code{mom_num_infect_dis}
    \item \code{mom_work_status}
    \item \code{mom_years_educ}
    \item \code{family_income}
    \item \code{housing_density}
    \item \code{mom_birth_place}
    \item \code{consanguinity}
    \item \code{socio_eco}
    \item \code{mom_race}
    \item \code{age_menarche}
    \item \code{dias_blood_pres}
    \item \code{mom_weight_birth}
    \item \code{dad_age}
    \item \code{dad_years_educ}
    \item \code{num_premes}
    \item \code{num_abortions}
    \item \code{num_prior_pregs}
    \item \code{num_stillbirths}
    \item \code{bayley_mental}
    \item \code{bayley_motor}
    \item \code{placental_weight}
    \item \code{cord_length}
    \item \code{sex}
    \item \code{apgar_1m_total}
    \item \code{apgar_5m_total}
    \item \code{bottle_feed_days}
    \item \code{breast_feed_days}
    \item \code{child_bilirubin}
    \item \code{child_hematocrit}
    \item \code{child_hemoglobin}
    \item \code{child_num_neur_abn}
    \item \code{child_num_cns_cond}
    \item \code{child_num_muscoskel}
    \item \code{child_num_resp_abn}
    \item \code{child_num_cardio_abn}
    \item \code{child_num_liver_abn}
    \item \code{child_num_hemo_cond}
    \item \code{child_num_infect}
    \item \code{child_num_synd}
    \item \code{child_num_endo_dis}
    \item \code{child_num_proc}
    \item \code{head_size_1yr}
    \item \code{gest_delivery}
  }
}
\source{
  Niswander, K. R. and Gordon, M. (1972) The Collaborative Perinatal Study of the National Institute of Neurological
  Diseases and Stroke: the women and their pregnancies. Philadelphia, PA: W.B. Saunders Company
  \url{https://www.archives.gov/research/electronic-records/nih.html}
}
\keyword{datasets}

