\name{input_2016}
\docType{data}
\alias{input_2016}
\title{Input Data for the 2016 ACIC Competition}
\description{
  Input data used in the 2016 Atlantic Causal Inference Competition, taken from the Collaborative Perinatal Project.
}
\usage{input_2016}
\format{A data frame consisting of 4802 observations and 58 covariates. The columns have been de-identified
        from their original source, but correspond to possible confounders, instruments, and uncorrelated
        variables from a hypothetical twin study on the impact of birthweight on IQ.}
\source{
  Niswander, K. R. and Gordon, M. (1972) The Collaborative Perinatal Study of the National Institute of Neurological
  Diseases and Stroke: the women and their pregnancies. Philadelphia, PA: W.B. Saunders Company
  \url{https://www.archives.gov/research/electronic-records/nih.html}
}
\keyword{datasets}

